Completed Get Edge Weight Function
